requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
mysqlclient==2.2.5
Flask-Mail==0.10.0
Authlib
Flask-OAuthlib
python-dotenv
